Move _Weighting to weighting_bases.py
