+++ src/torchjd/aggregation/imtl_g.py
@@ -3,7 +3,8 @@ from torch import Tensor
 
 from ._utils.gramian import compute_gramian
 from ._utils.non_differentiable import raise_non_differentiable_error
-from .bases import _WeightedAggregator, _Weighting
+from .bases import _WeightedAggregator
+from .weighting_bases import _Weighting
 
 
 class IMTLG(_WeightedAggregator):
