+++ src/torchjd/aggregation/random.py
@@ -2,7 +2,8 @@ import torch
 from torch import Tensor
 from torch.nn import functional as F
 
-from .bases import _WeightedAggregator, _Weighting
+from .bases import _WeightedAggregator
+from .weighting_bases import _Weighting
 
 
 class Random(_WeightedAggregator):
